openjdk integration
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@4718 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ core/src/openjdk/sun/sun/security/x509/EDIPartyName.java
@@ -38,7 +38,6 @@ import sun.security.util.*;
  * </pre>
  *
  * @author Hemma Prafullchandra
- * @version 1.18
  * @see GeneralName
  * @see GeneralNames
  * @see GeneralNameInterface
@@ -90,7 +89,7 @@ public class EDIPartyName implements GeneralNameInterface {
 
         for (int i = 0; i < len; i++) {
             DerValue opt = seq[i];
-            if (opt.isContextSpecific((byte)TAG_ASSIGNER) &&
+            if (opt.isContextSpecific(TAG_ASSIGNER) &&
                 !opt.isConstructed()) {
                 if (assigner != null)
                     throw new IOException("Duplicate nameAssigner found in"
@@ -98,7 +97,7 @@ public class EDIPartyName implements GeneralNameInterface {
                 opt = opt.data.getDerValue(); 
                 assigner = opt.getAsString();
             }
-            if (opt.isContextSpecific((byte)TAG_PARTYNAME) &&
+            if (opt.isContextSpecific(TAG_PARTYNAME) &&
                 !opt.isConstructed()) {
                 if (party != null)
                     throw new IOException("Duplicate partyName found in"
